Connect to stable hosts.
Previously it would connect to the host least-recently logged-into
and blacklist a host for two minutes after it lost its connection.

The blacklisting probably never did anything because of the least-recently
connected priority. Removed.

Order hosts by name then port, rather than randomly by HashMap key order
when (common case) they all have a last-login time of 0. But if a node
has had 3 errors in a row it sorts to the bottom.