Add updateBanned() override to allow adding a link.
A transparent and cheesy allowance for a subclass feature: massBan.
Maybe it should be imported here..

I wanted to break-up updateBanned() into subpieces that made sense
anyway, a setBanned() and setUnbanned() that each only took appropriate
arguments, even if the service call was wonky.

But it's a mess and screw it, let's just add a parameter at the end and
let subclassers be wary!

I suppose I can make the shorter-arg recordEvent()s final, since they
defer to longer-arg versions that can be overridden. That makes it safer
for overriders.